Code Encapsulation
Moved some part of code into a new file
